requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
cffi==1.14.0
cryptography==2.9
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
virtualenv==20.0.16
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.1.0
